Network-Training/requirements.txt
numpy==2.1.3
pandas==2.2.3
matplotlib==3.10.1
pillow==11.1.0
openpyxl==3.1.5
scikit-learn==1.6.1
tensorflow==2.19.0
ncps==1.0.1